Move stored functions into a separate file
